+++ pkg/rhttp/mux/router.go
@@ -21,8 +21,6 @@ package mux
 import (
 	"context"
 	"net/http"
-	"path"
-	"strings"
 )
 
 type Params map[string]string
@@ -61,23 +59,6 @@ type WalkFunc func(method, path string, handler http.Handler, opts *Options)
 
 type Walker interface {
 	Walk(ctx context.Context, f WalkFunc)
-}
-
-// ShiftPath splits off the first component of p, which will be cleaned of
-// relative components before processing. head will never contain a slash and
-// tail will always be a rooted path without trailing slash.
-// see https://blog.merovius.de/2017/06/18/how-not-to-use-an-http-router.html
-// and https://gist.github.com/weatherglass/62bd8a704d4dfdc608fe5c5cb5a6980c#gistcomment-2161690 for the zero alloc code below.
-func ShiftPath(p string) (head, tail string) {
-	if p == "" {
-		return "", "/"
-	}
-	p = strings.TrimPrefix(path.Clean(p), "/")
-	i := strings.Index(p, "/")
-	if i < 0 {
-		return p, "/"
-	}
-	return p[:i], p[i:]
 }
 
 func ParamsFromContext(ctx context.Context) Params {
